Small naming refactor, finish controllers, continue work on Spring Security

+++ src/main/java/com/skillstorm/taxguruplatform/repositories/Form1099Repository.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface Form1099Repository extends JpaRepository<Form1099, Integer> {
+public interface Form1099Repository extends JpaRepository<Form1099, Long> {
 
 }
